added-the-authorization-logics

+++ backend/routes/posts.js
@@ -33,8 +33,11 @@ router.post('',checkAuth,multer({storage:storage}).single('image'), (req, res, n
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + "/images/" + req.file.filename
+        imagePath: url + "/images/" + req.file.filename,
+        creator : req.userData.userId
     });
+    // console.log(req.userData)
+    // return res.status(200).json({})
     post.save().then((createdPost)=>{
         console.log(post);
         res.status(201).json({
@@ -97,14 +100,23 @@ router.put('/:id',checkAuth,multer({storage:storage}).single('image'),(req,res,n
             _id: req.body.id,
             title:req.body.title,
             content: req.body.content,
-            imagePath:imagePath
+            imagePath:imagePath,
+            creator : req.userData.userId
         })
         console.log(post)
-        Post.updateOne({_id:req.params.id},post).then((result)=>{
+        Post.updateOne({_id:req.params.id,creator:rq.userData.userId},post).then((result)=>{
             console.log(result);
-            res.status(200).json({
-                message: "Update Data Successfully"
-            })
+            if(result.nModified > 0 ){
+                res.status(200).json({
+                    message: "Update Data Successfully"
+                })
+            }
+            else{
+                res.status(401).json({
+                    message: "Not Authorized!"
+                })
+            }
+           
         })
 });
 
@@ -120,11 +132,19 @@ router.get('/:id',(req,res,next)=>{
 
 router.delete('/:id',checkAuth,(req,res,next)=>{
     console.log(req.params.id);
-    Post.deleteOne({_id:req.params.id}).then((result)=>{
+    Post.deleteOne({_id:req.params.id,creator:req.userData.userId}).then((result)=>{
         console.log(result);
-        res.status(200).json({
-            message: 'Sucessfully Deleted!!',
-        })
+        if(result.n > 0){
+            res.status(200).json({
+                message: 'Sucessfully Deleted!!',
+            })
+        }
+        else{
+            res.status(401).json({
+                message: 'Not Autharized!!',
+            })
+        }
+        
     });
     
 });
